bug fixes and improvements

+++ android/app/src/main/kotlin/com/ecency/waves/MainActivity.kt
@@ -43,7 +43,7 @@ class MainActivity: FlutterActivity() {
             val permlink = call.argument<String?>("permlink")
             val parentPermlink = call.argument<String?>("parentPermlink")
             val comment = call.argument<String?>("comment")
-            val weight = call.argument<Double?>("weight")
+            val weight = call.argument<Int?>("weight")
             if (id == null) {
                 result.error(
                     "UNAVAILABLE",
